+++ numerai_reports/reports.py
@@ -45,9 +45,9 @@ def payouts(models: List[str], groupby: str = "round") -> pd.DataFrame:
 
 def medals_leaderboard(limit: int = 10,
                        orderby: str = "total") -> pd.DataFrame:
-    cols = ['gold', 'silver', 'bronze']
-    lb = Data().leaderboard.groupby("model")[cols].sum()
-    lb['total'] = lb.sum(axis=1)
+    cols = ['gold', 'silver', 'bronze', 'active_rounds']
+    lb = Data().leaderboard[cols]
+    lb.insert(3, "total", lb[['gold', 'silver', 'bronze']].sum(axis=1))
     lb.sort_values(orderby, ascending=False, inplace=True)
     return lb.astype(int).fillna(0).head(limit)