+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## dev
+- add number of active rounds of a model to `medals_leaderboard`
+
 ## [0.5.3] - 2021-12-20
 - deprecate everything account related - information no longer provided my numerai
 - migrate to `V3UserProfile`